Fix compile on normal systems
This change fix the compile related to https://github.com/notspiff/screensavers.rsxs/pull/17,
where a cross compiler name was forced. On normal system it is mostly only "cc" and does not
work.

Before it brought:
```
configure: error: /bin/bash ./addons/build/screensavers.rsxs/lib/rsxs-1.0/config.sub cc failed
```

With this change it checks about them and add it only if it is not "cc".

The way how it comes in https://github.com/notspiff/screensavers.rsxs/pull/17 Is a bit of a
hack way, but not sure if there is a better way.